Remade everything since mobile became annoying, split into Map & Cards Components, mobile now working well
